Fixed the return type of the data coordinator responding to the number of iterations per epoch to match the unsigned type of the execution context.

+++ include/lbann/data_coordinator/data_coordinator.hpp
@@ -162,7 +162,7 @@ class data_coordinator {
   virtual const El::Matrix<El::Int>* get_sample_indices_per_mb(execution_mode mode) const = 0;
   virtual El::Matrix<El::Int>* get_sample_indices_per_mb(execution_mode mode) = 0;
 
-  virtual int get_num_iterations_per_epoch(execution_mode mode) const;
+  virtual size_t get_num_iterations_per_epoch(execution_mode mode) const;
 
   virtual int get_current_step_in_epoch(execution_mode mode) const;
